PythonForStm/requirements.txt
opencv-python
pyqt5==5.10
pyserial
